+++ sdk/src/index.ts
@@ -1,3 +1,3 @@
-export * from './shared';
-export * from './client';
-export * from './context';
+export * from './core/shared';
+export * from './core/client';
+export * from './core/context';
